backend/pyproject.toml
[project]
name = "makora-bio-backend"
version = "0.1.0"
description = "Flask backend for Makora Bio website with DSPY LLM workflows"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask",
    "flask-cors",
    "dspy",
    "python-dotenv",
    "anthropic>=0.71.0",
    "ruff>=0.14.2",
]

[project.optional-dependencies]
dev = [
    "jupyter",
    "ipykernel",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
